Revert unnecessary refactoring of frame.c
This fixes a crash when reading very large framed buffers.

rb_str_new takes a pointer to a string that it will memcpy into the
length sized char* buffer you're creating. This wouldn't fail all the time,
but it would at odd times when you had a very, very large string -- like a
JSON encoded a one million entry-sized array of integers.

Before we were new'ing up a string with rb_str_new and telling it to copy
from the buffer at the beginning of the index, as it was doing before.